added ft_split, ft_strdup, ft_strjoin nad ft_calloc

+++ ft_strlcat.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_strlcat(char *dst, const char *src, size_t size)
+size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
 	size_t i;
 	size_t src_len;
